# Betsy: Compress mipmaps in parallel