… modules/betsy/image_compress_betsy.h