Changes the Betsy implementation to compress mipmaps in parallel. This is meant to improve GPU utilization and speed up the compression process.
This was achieved by using 32 texture/image slots in each shader. This ensures that the reserved space is big enough for any supported texture (The maximal pixel count on each axis is 2^24).


TODO:
- [ ] Measure the performance difference,
- [ ] Test on low-end GPUs,
- [ ] Ensure the resulting textures look correct.