@fire 2025-01-05
I don't know if I can find the bandwidth to test, but these are my questions/investigations.

1. Does parallel betsy compiling work approximately the same as the old system in results?
2. Is it faster?
3. Is the code more complicated?
4. Since this code is load-bearing, I want to ensure it doesn't break anything since it's essential in the 3d import.

Other than that, the Betsy parallel compression looks good, and I want to get it in for testing, or we'll have to wait 6 months to a year for the next branch.