@BlueCube3310 2025-01-05
Imo it would be good to also get @clayjohn 's opinion on this, if possible