@BlueCube3310 2025-01-05
> I don't know if I can find the bandwidth to test, but these are my questions/investigations.
> 
>     1. Does parallel betsy compiling work approximately the same as the old system in results?
> 
>     2. Is it faster?
> 
>     3. Is the code more complicated?
> 
>     4. Since this code is load-bearing, I want to ensure it doesn't break anything since it's essential in the 3d import.

1. Yes, now the mipmaps are simply compressed at the same time, and after they're all done, they get copied from the GPU. Previously, each mipmap had to be compressed, then copied, one after another.

2. In 95% of cases they should be, since now the GPU is utilized more. On some systems with a GPU bottleneck it could be a bit slower, but I couldn't confirm it.

3. Kind of? In a way, it's simpler now, since previously it was kind of a mess with how the temporary textures were managed. 

4. I tested it on a few somewhat big projects, and it looks like it works as intended