@Calinou 2025-01-11
Tested locally, it works as expected.

However, I can't reproduce a performance improvement when importing large textures (a [4K PBR material](https://polyhaven.com/a/square_tiles_03) and a [8K4K HDRI](https://polyhaven.com/a/goegap_road), both with VRAM compression enabled).

## Benchmark

<details open>

<summary>PC specifications</summary>

- **CPU:** Intel Core i9-13900K
- **GPU:** NVIDIA GeForce RTX 4090
- **RAM:** 64 GB (2×32 GB DDR5-5800 C30)
- **SSD:** Solidigm P44 Pro 2 TB
- **OS:** Linux (Fedora 41)
</details>

File names with `hq` in them use the High Quality compresion option, which doesn't use Betsy for BPTC SDR. Betsy is used for BPTC HDR though. Either way, enabling High Quality has no effect on VRAM-compressed HDRIs, at least on desktop – they'll always use BPTC HDR.

### master

```
Betsy: Encoding a 4096x4096 image with 12 mipmaps as DXT1 RGB8 took 125 ms.
Betsy: Encoding a 4096x4096 image with 12 mipmaps as DXT1 RGB8 took 86 ms.
Betsy: Encoding a 4096x4096 image with 12 mipmaps as DXT1 RGB8 took 92 ms.
Betsy: Encoding a 4096x4096 image with 12 mipmaps as DXT1 RGB8 took 94 ms.
Betsy: Encoding a 8192x4096 image with 13 mipmaps as BPTC_RGBFU took 354 ms.
Betsy: Encoding a 8192x4096 image with 13 mipmaps as BPTC_RGBFU took 439 ms.
EditorFileSystem: "res://goegap_road_8k_hq.hdr" import took 2965 ms.
EditorFileSystem: "res://goegap_road_8k.hdr" import took 2972 ms.
EditorFileSystem: "res://textures_hq/square_tiles_03_arm_4k.jpg" import took 4891 ms.
EditorFileSystem: "res://textures_hq/square_tiles_03_diff_4k.jpg" import took 3902 ms.
EditorFileSystem: "res://textures_hq/square_tiles_03_disp_4k.jpg" import took 2018 ms.
EditorFileSystem: "res://textures_hq/square_tiles_03_nor_gl_4k.jpg" import took 5763 ms.
EditorFileSystem: "res://textures/square_tiles_03_arm_4k.jpg" import took 2790 ms.
EditorFileSystem: "res://textures/square_tiles_03_diff_4k.jpg" import took 2794 ms.
EditorFileSystem: "res://textures/square_tiles_03_disp_4k.jpg" import took 1700 ms.
EditorFileSystem: "res://textures/square_tiles_03_nor_gl_4k.jpg" import took 2793 ms.

```

### This PR

```
Betsy: Encoding a 4096x4096 image with 12 mipmaps as DXT1 RGB8 took 111 ms.
Betsy: Encoding a 4096x4096 image with 12 mipmaps as DXT1 RGB8 took 114 ms.
Betsy: Encoding a 4096x4096 image with 12 mipmaps as DXT1 RGB8 took 131 ms.
Betsy: Encoding a 4096x4096 image with 12 mipmaps as DXT1 RGB8 took 218 ms.
Betsy: Encoding a 8192x4096 image with 13 mipmaps as BPTC_RGBFU took 488 ms.
Betsy: Encoding a 8192x4096 image with 13 mipmaps as BPTC_RGBFU took 495 ms.
EditorFileSystem: "res://goegap_road_8k_hq.hdr" import took 3364 ms.
EditorFileSystem: "res://goegap_road_8k.hdr" import took 3371 ms.
EditorFileSystem: "res://textures_hq/square_tiles_03_arm_4k.jpg" import took 5840 ms.
EditorFileSystem: "res://textures_hq/square_tiles_03_diff_4k.jpg" import took 3843 ms.
EditorFileSystem: "res://textures_hq/square_tiles_03_disp_4k.jpg" import took 1964 ms.
EditorFileSystem: "res://textures_hq/square_tiles_03_nor_gl_4k.jpg" import took 5081 ms.
EditorFileSystem: "res://textures/square_tiles_03_arm_4k.jpg" import took 3186 ms.
EditorFileSystem: "res://textures/square_tiles_03_diff_4k.jpg" import took 3190 ms.
EditorFileSystem: "res://textures/square_tiles_03_disp_4k.jpg" import took 1829 ms.
EditorFileSystem: "res://textures/square_tiles_03_nor_gl_4k.jpg" import took 3190 ms.
```